Large rewrite to address bugs and new yaml file
I've changed the format of the yaml file to accommodate multiple cloud vendors.
This script is exclusive for AWS instances, however the yaml file can be referenced
by other scripts to create servers on Digital Ocean and Linode.

Maybe a larger project is to consolidate all of the scripts into a single codebase.